project/requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
azure-cosmos==4.5.1
openai==0.28.1
requests==2.31.0
reportlab==4.0.4
docx2txt==0.8
PyPDF2==3.0.1
azure-cognitiveservices-speech==1.30.0